Function updated

+++ elasticsearch_project/src/main/java/org/snp/service/data/DataService.java
@@ -23,6 +23,9 @@ public class DataService {
 
     @Inject FileService fileService;
 
+    final String NODE_NAME = Main.isMasterTest() ? "Master" : System.getProperty("name");
+    final String MESSAGE_PREFIX = NODE_NAME + " : ";
+
     private TableDao tableDao = new TableDao();
     private DataDao dataDAO = new DataDao();
 
@@ -41,7 +44,7 @@ public class DataService {
     public Message query(QueryCredentials queryCredentials){
         Table table = tableDao.find(queryCredentials.tableName);
         if(table == null)
-            return new MessageAttachment<>(404, "table "+ queryCredentials.tableName+" does not exists");
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"table "+ queryCredentials.tableName+" does not exists");
 
         List<String> references;
         /**
@@ -52,40 +55,44 @@ public class DataService {
             for (QueryCredentials.AttributeCredentials attributeCredentials : queryCredentials.queryParams) {
                 String columnName = attributeCredentials.columnName;
                 if (!table.containsColumn(columnName))
-                    return new MessageAttachment<>(404, "column " + columnName + " does not exists in " + queryCredentials.tableName);
+                    return new MessageAttachment<>(404, MESSAGE_PREFIX+"column " + columnName + " does not exists in " + queryCredentials.tableName);
                 queryMap.put(attributeCredentials.columnName, attributeCredentials.value);
             }
 
 
             references = dataDAO.find(table, queryMap);
-            if (references == null || references.isEmpty())
-                return new MessageAttachment<>(404, "data not found");
+            if (!Main.isMasterTest() && (references == null || references.isEmpty()))
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
         }else{
             references = dataDAO.findAll(table);
         }
-        /**
-         * all the row
-         */
+
         List<String> values = new ArrayList<>();
-        for(String ref : references){
-            String[] refSplited = ref.split(",");
-            values.add(fileService.getAllDataAtPos(refSplited[0], Integer.valueOf(refSplited[1]) ));
-        }
 
-        /**
-         * column selected verification
-         */
-        if(queryCredentials.columnsSelected!=null) {
-            List<String> columnsName = new ArrayList<>();
-            for (ColumnCredentials columnCredentials : queryCredentials.columnsSelected) {
-                String columnName = columnCredentials.name;
-                if (!table.containsColumn(columnName))
-                    return new MessageAttachment<>(404, "column " + columnName + " does not exists in " + queryCredentials.tableName);
-                columnsName.add(columnName);
+        boolean doNotContinueTest = references == null || references.isEmpty();
+        if(! doNotContinueTest) {
+            /**
+             * all the row
+             */
+            for (String ref : references) {
+                String[] refSplited = ref.split(",");
+                values.add(fileService.getAllDataAtPos(refSplited[0], Integer.valueOf(refSplited[1])));
             }
+            /**
+             * column selected verification
+             */
+            if (queryCredentials.columnsSelected != null) {
+                List<String> columnsName = new ArrayList<>();
+                for (ColumnCredentials columnCredentials : queryCredentials.columnsSelected) {
+                    String columnName = columnCredentials.name;
+                    if (!table.containsColumn(columnName))
+                        return new MessageAttachment<>(404, MESSAGE_PREFIX + "column " + columnName + " does not exists in " + queryCredentials.tableName);
+                    columnsName.add(columnName);
+                }
 
-            values = getValuesForColumn(table, columnsName, values);
+                values = getValuesForColumn(table, columnsName, values);
 
+            }
         }
 
         /**
@@ -98,6 +105,8 @@ public class DataService {
                 if(slaveResult!=null)
                     values.addAll(slaveResult);
             }
+            if(values.isEmpty())
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
         }
          return new MessageAttachment<List>(200, values);
     }
@@ -110,7 +119,7 @@ public class DataService {
     public Message delete(QueryCredentials queryCredentials){
         Table table = tableDao.find(queryCredentials.tableName);
         if(table == null)
-            return new MessageAttachment<>(404, "table "+ queryCredentials.tableName+" does not exists");
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"table "+ queryCredentials.tableName+" does not exists");
 
         List<String> references;
         /**
@@ -121,24 +130,27 @@ public class DataService {
             for (QueryCredentials.AttributeCredentials attributeCredentials : queryCredentials.queryParams) {
                 String columnName = attributeCredentials.columnName;
                 if (!table.containsColumn(columnName))
-                    return new MessageAttachment<>(404, "column " + columnName + " does not exists in " + queryCredentials.tableName);
+                    return new MessageAttachment<>(404, MESSAGE_PREFIX+"column " + columnName + " does not exists in " + queryCredentials.tableName);
                 queryMap.put(attributeCredentials.columnName, attributeCredentials.value);
             }
 
 
             references = dataDAO.find(table, queryMap);
-            if (references == null || references.isEmpty())
-                return new MessageAttachment<>(404, "data not found");
+            if (!Main.isMasterTest() && (references == null || references.isEmpty()))
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
         }else{
             references = dataDAO.findAll(table);
         }
 
-        /**
-         * suppression in subindex
-         */
-        for(SubIndex subIndex : table.getSubIndexMap().values()){
-            for(String ref : references){
-                subIndex.deleteByReference(ref);
+        boolean doNotContinueTest = references == null || references.isEmpty();
+        if(! doNotContinueTest) {
+            /**
+             * suppression in subindex
+             */
+            for (SubIndex subIndex : table.getSubIndexMap().values()) {
+                for (String ref : references) {
+                    subIndex.deleteByReference(ref);
+                }
             }
         }
 
@@ -150,6 +162,9 @@ public class DataService {
             for(SlaveClient slaveClient : slaveClients){
                 finalResult += slaveClient.dataDelete(queryCredentials);
             }
+            if(finalResult==0)
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
+
         }
 
         return new MessageAttachment<>(200, finalResult);
@@ -159,15 +174,14 @@ public class DataService {
      * UPDATE
      *
      */
-
     public Message update(QueryCredentials queryCredentials){
         Table table = tableDao.find(queryCredentials.tableName);
         if(table == null)
-            return new MessageAttachment<>(404, "table "+ queryCredentials.tableName+" does not exists");
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"table "+ queryCredentials.tableName+" does not exists");
         if(queryCredentials.updateParams == null)
-            return new MessageAttachment<>(404, "update_params could not be null");
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"update_params could not be null");
         if(queryCredentials.queryParams==null)
-            return new MessageAttachment<>(404, "query_params could not be null");
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"query_params could not be null");
 
         List<String> references;
         /**
@@ -178,59 +192,65 @@ public class DataService {
         for (QueryCredentials.AttributeCredentials attributeCredentials : queryCredentials.queryParams) {
             String columnName = attributeCredentials.columnName;
             if (!table.containsColumn(columnName))
-                return new MessageAttachment<>(404, "column " + columnName + " does not exists in " + queryCredentials.tableName);
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"column " + columnName + " does not exists in " + queryCredentials.tableName);
             queryMap.put(attributeCredentials.columnName, attributeCredentials.value);
         }
 
 
         references = dataDAO.find(table, queryMap);
-        if (references == null || references.isEmpty())
-            return new MessageAttachment<>(404, "data not found");
+        if (!Main.isMasterTest() && (references == null || references.isEmpty()))
+            return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
 
+        boolean doNotContinueTest = references == null || references.isEmpty();
+        int finalResult=0;
+        if(! doNotContinueTest) {
+            /**
+             * update
+             */
+            List<String> values = new ArrayList<>();
+            for (Map.Entry entry : table.getSubIndexMap().entrySet()) {
+                for (QueryCredentials.AttributeCredentials attributeCredentials : queryCredentials.updateParams) {
+                    String columnName = attributeCredentials.columnName;
+                    int columnNumber = table.positionOfColumn(columnName);
+                    String newValue = attributeCredentials.value;
 
-        /**
-         * update
-         */
-        List<String> values = new ArrayList<>();
-        for(Map.Entry entry : table.getSubIndexMap().entrySet()){
-            for (QueryCredentials.AttributeCredentials attributeCredentials : queryCredentials.updateParams){
-                String columnName = attributeCredentials.columnName;
-                int columnNumber = table.positionOfColumn(columnName);
-                String newValue = attributeCredentials.value;
-
-                if(columnName.equals(entry.getKey())){
-                    SubIndex subIndex = (SubIndex) entry.getValue();
-                    int size = references.size();
-                    for(int i = 0; i<references.size(); ){
-                        String ref = references.get(i);
-                        /**
-                         * update in subindex
-                         */
-                        subIndex.updateByReference(newValue, ref);
-                        /**
-                         * update in file
-                         */
-                        String[] refSplited = ref.split(",");
-                        values.add(fileService.updateColumnAtPos(table, refSplited[0], Integer.valueOf(refSplited[1]), columnNumber, newValue));
-                        //update the position byte
-                        references = dataDAO.find(table, queryMap);
-                        if (references == null || references.isEmpty())
-                            break;
-                        if(references.size() == size)
-                            i++;
+                    if (columnName.equals(entry.getKey())) {
+                        SubIndex subIndex = (SubIndex) entry.getValue();
+                        int size = references.size();
+                        finalResult = size;
+                        for (int i = 0; i < references.size(); ) {
+                            String ref = references.get(i);
+                            /**
+                             * update in subindex
+                             */
+                            subIndex.updateByReference(newValue, ref);
+                            /**
+                             * update in file
+                             */
+                            String[] refSplited = ref.split(",");
+                            values.add(fileService.updateColumnAtPos(table, refSplited[0], Integer.valueOf(refSplited[1]), columnNumber, newValue));
+                            //cas où on modifie ce qu'on recherche : exemple update nom=Y where nom=X
+                            references = dataDAO.find(table, queryMap);
+                            if (references == null || references.isEmpty())
+                                break;
+                            if (references.size() == size)
+                                i++;
+                        }
                     }
                 }
             }
         }
 
-        int finalResult = references.size();
         /**
          * redirection to the slaves
          */
+
         if(Main.isMasterTest()){
             for(SlaveClient slaveClient : slaveClients){
                 finalResult += slaveClient.dataUpdate(queryCredentials);
             }
+            if(finalResult==0)
+                return new MessageAttachment<>(404, MESSAGE_PREFIX+"data not found");
         }
